writing-tests/lint-tool.html

<!DOCTYPE html>

<html lang="en" data-content_root="../">
  <head>
    <meta charset="utf-8" />
    <meta name="viewport" content="width=device-width, initial-scale=1.0" /><meta name="viewport" content="width=device-width, initial-scale=1" />

    <title>Lint Tool &#8212; web-platform-tests  documentation</title>
    <link rel="stylesheet" type="text/css" href="../_static/pygments.css?v=5ecbeea2" />
    <link rel="stylesheet" type="text/css" href="../_static/alabaster.css?v=12dfc556" />
    <script src="../_static/documentation_options.js?v=5929fcd5"></script>
    <script src="../_static/doctools.js?v=888ff710"></script>
    <script src="../_static/sphinx_highlight.js?v=dc90522c"></script>
    <link rel="index" title="Index" href="../genindex.html" />
    <link rel="search" title="Search" href="../search.html" />
    <link rel="next" title="The Ahem Font" href="ahem.html" />
    <link rel="prev" title="Submitting Tests" href="submission-process.html" />
   
  <link rel="stylesheet" href="../_static/custom.css" type="text/css" />
  

  
  

  </head><body>
  

    <div class="document">
      <div class="documentwrapper">
        <div class="bodywrapper">
          

          <div class="body" role="main">
            
  <section id="lint-tool">
<h1>Lint Tool<a class="headerlink" href="#lint-tool" title="Link to this heading">¶</a></h1>
<p>We have a lint tool for catching common mistakes in test files. You can run
it manually by running the <code class="docutils literal notranslate"><span class="pre">wpt</span> <span class="pre">lint</span></code> command from the root of your local
web-platform-tests working directory like this:</p>
<div class="highlight-default notranslate"><div class="highlight"><pre><span></span><span class="o">./</span><span class="n">wpt</span> <span class="n">lint</span>
</pre></div>
</div>
<p>The lint tool is also run automatically for every submitted pull request,
and reviewers will not merge branches with tests that have lint errors, so
you must either <a class="reference external" href="#fixing-lint-errors">fix all lint errors</a>, or you must
<a class="reference external" href="#updating-the-ignored-files">add an exception</a> to suppress the errors.</p>
<section id="fixing-lint-errors">
<h2>Fixing lint errors<a class="headerlink" href="#fixing-lint-errors" title="Link to this heading">¶</a></h2>
<p>You must fix any errors the lint tool reports, unless an error is for something
essential to a certain test or that for some other exceptional reason shouldn’t
prevent the test from being merged; in those cases you can <a class="reference external" href="#updating-the-ignored-files">add an
exception</a> to suppress the errors. In all other
cases, follow the instructions below to fix all errors reported.</p>
<!--
  This listing is automatically generated from the linting tool's Python source
  code.
--><dl>
<dt><code class="docutils literal notranslate"><span class="pre">ABSOLUTE-URL-REF</span></code></dt><dd><p>Reference test with a reference file specified via an absolute URL: ‘%s’</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">AHEM</span> <span class="pre">COPY</span></code></dt><dd><p>Don’t add extra copies of Ahem, use /fonts/Ahem.ttf</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">AHEM</span> <span class="pre">SYSTEM</span> <span class="pre">FONT</span></code></dt><dd><p>Don’t use Ahem as a system font, use /fonts/ahem.css</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">ASSERT-PRECONDITION</span></code></dt><dd><p>Test-file line has an <code class="docutils literal notranslate"><span class="pre">assert_precondition(...)</span></code> call</p>
<strong>To fix:</strong><p>Replace with <code class="docutils literal notranslate"><span class="pre">assert_implements</span></code> or <code class="docutils literal notranslate"><span class="pre">assert_implements_optional</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">ASSERT_THROWS</span></code></dt><dd><p>Test-file line has an <code class="docutils literal notranslate"><span class="pre">assert_throws(...)</span></code> call</p>
<strong>To fix:</strong><p>Replace with <code class="docutils literal notranslate"><span class="pre">assert_throws_dom</span></code> or <code class="docutils literal notranslate"><span class="pre">assert_throws_js</span></code> or <code class="docutils literal notranslate"><span class="pre">assert_throws_exactly</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">BROKEN-GLOBAL-METADATA</span></code></dt><dd><p>Invalid global metadata: %s</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">BROKEN-METADATA</span></code></dt><dd><p>Metadata comment is not formatted correctly</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">CONSOLE</span></code></dt><dd><p>Test-file line has a <code class="docutils literal notranslate"><span class="pre">console.*(...)</span></code> call</p>
<strong>To fix:</strong><p>remove the <code class="docutils literal notranslate"><span class="pre">console.*(...)</span></code> call (and in some cases, consider adding an
<code class="docutils literal notranslate"><span class="pre">assert_*</span></code> of some kind in place of it)</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">CONTENT-MANUAL</span></code></dt><dd><p>Manual test whose filename doesn’t end in ‘-manual’</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">CONTENT-VISUAL</span></code></dt><dd><p>Visual test whose filename doesn’t end in ‘-visual’</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">CR</span> <span class="pre">AT</span> <span class="pre">EOL</span></code></dt><dd><p>Test-file line ends with CR (U+000D) character</p>
<strong>To fix:</strong><p>reformat file so each line just has LF (U+000A) line ending (standard,
cross-platform “Unix” line endings instead of, e.g., DOS line endings).</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">DUPLICATE-BASENAME-PATH</span></code></dt><dd><p>File has identical basename path (path excluding extension) as other file(s) (found extensions: %s)</p>
<strong>To fix:</strong><p>rename files so they have unique basename paths</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">DUPLICATE-CASE-INSENSITIVE-PATH</span></code></dt><dd><p>Path differs from path %s only in case</p>
<strong>To fix:</strong><p>rename files so they are unique irrespective of case</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">EARLY-TESTDRIVER-VENDOR</span></code></dt><dd><p>Test file has an instance of <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testdriver-vendor.js'&gt;</span></code> prior to <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testdriver.js'&gt;</span></code></p>
<strong>To fix:</strong><p>flip the order</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">EARLY-TESTHARNESSREPORT</span></code></dt><dd><p>Test file has an instance of <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharnessreport.js'&gt;</span></code> prior to <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharness.js'&gt;</span></code></p>
<strong>To fix:</strong><p>flip the order</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">FILE</span> <span class="pre">TYPE</span></code></dt><dd><p>/%s is an unsupported file type (%s)</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">GENERATE_TESTS</span></code></dt><dd><p>Test file line has a generate_tests call</p>
<strong>To fix:</strong><p>remove the call and call <code class="docutils literal notranslate"><span class="pre">test()</span></code> a number of times instead</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">GITIGNORE</span></code></dt><dd><p>.gitignore found outside the root</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">HTML</span> <span class="pre">INVALID</span> <span class="pre">SYNTAX</span></code></dt><dd><p>Test-file line has a non-void HTML tag with /&gt; syntax</p>
<strong>To fix:</strong><p>Replace with start tag and end tag</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">IGNORED</span> <span class="pre">PATH</span></code></dt><dd><p>%s matches an ignore filter in .gitignore - please add a .gitignore exception</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">INDENT</span> <span class="pre">TABS</span></code></dt><dd><p>Test-file line starts with one or more tab characters</p>
<strong>To fix:</strong><p>use spaces to replace any tab characters at beginning of lines</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">INDENTED-METADATA</span></code></dt><dd><p>Metadata comments should start the line</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">INVALID-META-FILE</span></code></dt><dd><p>The META.yml is not a YAML file with the expected structure</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">INVALID-TIMEOUT</span></code></dt><dd><p>Test file with <code class="docutils literal notranslate"><span class="pre">&lt;meta</span> <span class="pre">name='timeout'...&gt;</span></code> element that has a <code class="docutils literal notranslate"><span class="pre">content</span></code> attribute whose value is not <code class="docutils literal notranslate"><span class="pre">long</span></code>: %s</p>
<strong>To fix:</strong><p>replace the value of the <code class="docutils literal notranslate"><span class="pre">content</span></code> attribute with <code class="docutils literal notranslate"><span class="pre">long</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">INVALID-WEB-FEATURES-FILE</span></code></dt><dd><p>The WEB_FEATURES.yml file contains an invalid structure</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">LATE-TIMEOUT</span></code></dt><dd><p>Test file with <code class="docutils literal notranslate"><span class="pre">&lt;meta</span> <span class="pre">name='timeout'...&gt;</span></code> element after <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharnessreport.js'&gt;</span></code> element</p>
<strong>To fix:</strong><p>move the <code class="docutils literal notranslate"><span class="pre">&lt;meta</span> <span class="pre">name=&quot;timeout&quot;...&gt;</span></code> element to precede the <code class="docutils literal notranslate"><span class="pre">script</span></code>
element.</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">LAYOUTTESTS</span> <span class="pre">APIS</span></code></dt><dd><p>eventSender/testRunner/internals used; these are LayoutTests-specific APIs (WebKit/Blink)</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MALFORMED-VARIANT</span></code></dt><dd><p>%s must be a non empty string and start with ‘?’ or ‘#’</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MISSING</span> <span class="pre">DEPENDENCY</span></code></dt><dd><p>Chromium-specific content referenced</p>
<strong>To fix:</strong><p>Reimplement the test to use well-documented testing interfaces</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MISSING-LINK</span></code></dt><dd><p>Testcase file must have a link to a spec</p>
<strong>To fix:</strong><p>Ensure that there is a <code class="docutils literal notranslate"><span class="pre">&lt;link</span> <span class="pre">rel=&quot;help&quot;</span> <span class="pre">href=&quot;[url]&quot;&gt;</span></code> for the spec.
<code class="docutils literal notranslate"><span class="pre">MISSING-LINK</span></code> is designed to ensure that the CSS build tool can find
the tests. Note that the CSS build system is primarily used by
<a class="reference external" href="http://test.csswg.org/">test.csswg.org/</a>, which doesn’t use
<code class="docutils literal notranslate"><span class="pre">wptserve</span></code>, so <code class="docutils literal notranslate"><span class="pre">*.any.js</span></code> and similar tests won’t work there; stick
with the <code class="docutils literal notranslate"><span class="pre">.html</span></code> equivalent.</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MISSING-REFTESTWAIT</span></code></dt><dd><p>Missing <code class="docutils literal notranslate"><span class="pre">class=reftest-wait</span></code></p>
<strong>To fix:</strong><p>ensure tests that include reftest-wait.js also use class=reftest-wait on the root element.</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MISSING-TESTDRIVER-VENDOR</span></code></dt><dd><p>Missing <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testdriver-vendor.js'&gt;</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MISSING-TESTHARNESSREPORT</span></code></dt><dd><p>Missing <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharnessreport.js'&gt;</span></code></p>
<strong>To fix:</strong><p>ensure each test file contains <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharnessreport.js'&gt;</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MISSING-WEB-FEATURES-FILE</span></code></dt><dd><p>The WEB_FEATURES.yml file references a test that does not exist: ‘%s’</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MOJOM-JS</span></code></dt><dd><p>Don’t check *.mojom.js files into WPT</p>
<strong>To fix:</strong><p>Check if the file is already included in mojojs.zip:
https://source.chromium.org/chromium/chromium/src/+/master:chrome/tools/build/linux/FILES.cfg
If yes, use <code class="docutils literal notranslate"><span class="pre">loadMojoResources</span></code> from <code class="docutils literal notranslate"><span class="pre">resources/test-only-api.js</span></code> to load
it; if not, contact ecosystem-infra&#64;chromium.org for adding new files
to mojojs.zip.</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MULTIPLE-TESTDRIVER</span></code></dt><dd><p>More than one <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testdriver.js'&gt;</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MULTIPLE-TESTDRIVER-VENDOR</span></code></dt><dd><p>More than one <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testdriver-vendor.js'&gt;</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MULTIPLE-TESTHARNESS</span></code></dt><dd><p>More than one <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharness.js'&gt;</span></code></p>
<strong>To fix:</strong><p>Ensure each test has only one <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharness.js'&gt;</span></code> instance.
For <code class="docutils literal notranslate"><span class="pre">.js</span></code> tests, remove <code class="docutils literal notranslate"><span class="pre">//</span> <span class="pre">META:</span> <span class="pre">script=/resources/testharness.js</span></code>,
which wptserve already adds to the boilerplate markup.</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MULTIPLE-TESTHARNESSREPORT</span></code></dt><dd><p>More than one <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharnessreport.js'&gt;</span></code></p>
<strong>To fix:</strong><p>Ensure each test has only one <code class="docutils literal notranslate"><span class="pre">&lt;script</span> <span class="pre">src='/resources/testharnessreport.js'&gt;</span></code> instance.
For <code class="docutils literal notranslate"><span class="pre">.js</span></code> tests, remove <code class="docutils literal notranslate"><span class="pre">//</span> <span class="pre">META:</span> <span class="pre">script=/resources/testharnessreport.js</span></code>,
which wptserve already adds to the boilerplate markup.</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">MULTIPLE-TIMEOUT</span></code></dt><dd><p>More than one meta name=’timeout’</p>
<strong>To fix:</strong><p>ensure each test file has only one instance of a <code class="docutils literal notranslate"><span class="pre">&lt;meta</span> <span class="pre">name=&quot;timeout&quot;...&gt;</span></code> element</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">NON-EXISTENT-REF</span></code></dt><dd><p>Reference test with a non-existent ‘%s’ relationship reference: ‘%s’</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">OPEN-NO-MODE</span></code></dt><dd><p>File opened without providing an explicit mode (note: binary files must be read with ‘b’ in the mode flags)</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">PARSE-FAILED</span></code></dt><dd><p>Unable to parse file</p>
<strong>To fix:</strong><p>examine the file to find the causes of any parse errors, and fix them.</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">PATH</span> <span class="pre">LENGTH</span></code></dt><dd><p>/%s longer than maximum path length (%d &gt; 150)</p>
<strong>To fix:</strong><p>use shorter filename to rename the test file</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">PRINT</span> <span class="pre">STATEMENT</span></code></dt><dd><p>A server-side python support file contains a <code class="docutils literal notranslate"><span class="pre">print</span></code> statement</p>
<strong>To fix:</strong><p>remove the <code class="docutils literal notranslate"><span class="pre">print</span></code> statement or replace it with something else that
achieves the intended effect (e.g., a logging call)</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">PROMISE_REJECTS</span></code></dt><dd><p>Test-file line has a <code class="docutils literal notranslate"><span class="pre">promise_rejects(...)</span></code> call</p>
<strong>To fix:</strong><p>Replace with promise_rejects_dom or promise_rejects_js or <code class="docutils literal notranslate"><span class="pre">promise_rejects_exactly</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">REFERENCE-IN-OTHER-TYPE</span></code></dt><dd><p>Reference link included in a %s test</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">SAME-FILE-REF</span></code></dt><dd><p>Reference test which points at itself as a reference</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">SET</span> <span class="pre">TIMEOUT</span></code></dt><dd><p>setTimeout used</p>
<strong>To fix:</strong><p>replace all <code class="docutils literal notranslate"><span class="pre">setTimeout(...)</span></code> calls with <code class="docutils literal notranslate"><span class="pre">step_timeout(...)</span></code> calls</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">SPECIALPOWERS</span> <span class="pre">API</span></code></dt><dd><p>SpecialPowers used; this is gecko-specific and not supported in wpt</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">STRAY-METADATA</span></code></dt><dd><p>Metadata comments should start the file</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TENTATIVE-DIRECTORY-NAME</span></code></dt><dd><p>Directories for tentative tests must be named exactly ‘tentative’</p>
<strong>To fix:</strong><p>rename directory to be called ‘tentative’</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TEST</span> <span class="pre">DRIVER</span> <span class="pre">INTERNAL</span></code></dt><dd><p>Test-file uses test_driver_internal API</p>
<strong>To fix:</strong><p>Only use test_driver public API</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TESTDRIVER-IN-UNSUPPORTED-TYPE</span></code></dt><dd><p>testdriver.js included in a %s test, which doesn’t support testdriver.js</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TESTDRIVER-PATH</span></code></dt><dd><p>testdriver.js script seen with incorrect path</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TESTDRIVER-UNSUPPORTED-QUERY-PARAMETER</span></code></dt><dd><p>testdriver.js script seen with unsupported query parameters</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TESTDRIVER-VENDOR-PATH</span></code></dt><dd><p>testdriver-vendor.js script seen with incorrect path</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TESTHARNESS-IN-OTHER-TYPE</span></code></dt><dd><p>testharness.js included in a %s test</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TESTHARNESS-PATH</span></code></dt><dd><p>testharness.js script seen with incorrect path</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TESTHARNESSREPORT-PATH</span></code></dt><dd><p>testharnessreport.js script seen with incorrect path</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">TRAILING</span> <span class="pre">WHITESPACE</span></code></dt><dd><p>Whitespace at EOL</p>
<strong>To fix:</strong><p>Remove trailing whitespace from all lines in the file.</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">UNKNOWN-GLOBAL-METADATA</span></code></dt><dd><p>Unexpected value for global metadata</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">UNKNOWN-METADATA</span></code></dt><dd><p>Unexpected kind of metadata</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">UNKNOWN-TIMEOUT-METADATA</span></code></dt><dd><p>Unexpected value for timeout metadata</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">VARIANT-MISSING</span></code></dt><dd><p>Test file with a <code class="docutils literal notranslate"><span class="pre">&lt;meta</span> <span class="pre">name='variant'...&gt;</span></code> element that’s missing a <code class="docutils literal notranslate"><span class="pre">content</span></code> attribute</p>
<strong>To fix:</strong><p>add a <code class="docutils literal notranslate"><span class="pre">content</span></code> attribute with an appropriate value to the <code class="docutils literal notranslate"><span class="pre">&lt;meta</span> <span class="pre">name='variant'...&gt;</span></code> element</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">W3C-TEST.ORG</span></code></dt><dd><p>Test-file line has the string <code class="docutils literal notranslate"><span class="pre">w3c-test.org</span></code></p>
<strong>To fix:</strong><p>either replace the <code class="docutils literal notranslate"><span class="pre">w3c-test.org</span></code> string with the expression
<code class="docutils literal notranslate"><span class="pre">{{host}}:{{ports[http][0]}}</span></code> or a generic hostname like <code class="docutils literal notranslate"><span class="pre">example.org</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">WEB-PLATFORM.TEST</span></code></dt><dd><p>Internal web-platform.test domain used</p>
<strong>To fix:</strong><p>use <a class="reference external" href="https://web-platform-tests.org/writing-tests/server-pipes.html#sub">server-side substitution</a>,
along with the <a class="reference external" href="https://web-platform-tests.org/writing-tests/file-names.html#test-features"><code class="docutils literal notranslate"><span class="pre">.sub</span></code> filename-flag</a>,
to replace web-platform.test with <code class="docutils literal notranslate"><span class="pre">{{domains[]}}</span></code></p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">WEBIDL2.JS</span></code></dt><dd><p>Legacy webidl2.js script used</p>
</dd>
<dt><code class="docutils literal notranslate"><span class="pre">WORKER</span> <span class="pre">COLLISION</span></code></dt><dd><p>path ends with %s which collides with generated tests from %s files</p>
</dd>
</dl>
</section>
<section id="updating-the-ignored-files">
<h2>Updating the ignored files<a class="headerlink" href="#updating-the-ignored-files" title="Link to this heading">¶</a></h2>
<p>Normally you must <a class="reference external" href="#fixing-lint-errors">fix all lint errors</a>. But in the
unusual case of error reports for things essential to certain tests or that
for other exceptional reasons shouldn’t prevent a merge of a test, you can
update and commit the <code class="docutils literal notranslate"><span class="pre">lint.ignore</span></code> file in the web-platform-tests root
directory to suppress errors the lint tool would report for a test file.</p>
<p>To add a test file or directory to the list, use the following format:</p>
<div class="highlight-default notranslate"><div class="highlight"><pre><span></span><span class="n">ERROR</span> <span class="n">TYPE</span><span class="p">:</span><span class="n">file</span><span class="o">/</span><span class="n">name</span><span class="o">/</span><span class="n">pattern</span>
</pre></div>
</div>
<p>For example, to ignore all <code class="docutils literal notranslate"><span class="pre">TRAILING</span> <span class="pre">WHITESPACE</span></code> errors in the file
<code class="docutils literal notranslate"><span class="pre">example/file.html</span></code>, add the following line to the <code class="docutils literal notranslate"><span class="pre">lint.ignore</span></code> file:</p>
<div class="highlight-default notranslate"><div class="highlight"><pre><span></span><span class="n">TRAILING</span> <span class="n">WHITESPACE</span><span class="p">:</span><span class="n">example</span><span class="o">/</span><span class="n">file</span><span class="o">.</span><span class="n">html</span>
</pre></div>
</div>
<p>To ignore errors for an entire directory rather than just one file, use the <code class="docutils literal notranslate"><span class="pre">*</span></code>
wildcard. For example, to ignore all <code class="docutils literal notranslate"><span class="pre">TRAILING</span> <span class="pre">WHITESPACE</span></code> errors in the
<code class="docutils literal notranslate"><span class="pre">example</span></code> directory, add the following line to the <code class="docutils literal notranslate"><span class="pre">lint.ignore</span></code> file:</p>
<div class="highlight-default notranslate"><div class="highlight"><pre><span></span><span class="n">TRAILING</span> <span class="n">WHITESPACE</span><span class="p">:</span><span class="n">example</span><span class="o">/*</span>
</pre></div>
</div>
<p>Similarly, you can also
use
<a class="reference external" href="https://docs.python.org/library/fnmatch.html">shell-style wildcards</a> to
express other filename patterns or directory-name patterns.</p>
<p>Finally, to ignore just one line in a file, use the following format:</p>
<div class="highlight-default notranslate"><div class="highlight"><pre><span></span><span class="n">ERROR</span> <span class="n">TYPE</span><span class="p">:</span><span class="n">file</span><span class="o">/</span><span class="n">name</span><span class="o">/</span><span class="n">pattern</span><span class="p">:</span><span class="n">line_number</span>
</pre></div>
</div>
<p>For example, to ignore the <code class="docutils literal notranslate"><span class="pre">TRAILING</span> <span class="pre">WHITESPACE</span></code> error for just line 128 of the
file <code class="docutils literal notranslate"><span class="pre">example/file.html</span></code>, add the following to the <code class="docutils literal notranslate"><span class="pre">lint.ignore</span></code> file:</p>
<div class="highlight-default notranslate"><div class="highlight"><pre><span></span><span class="n">TRAILING</span> <span class="n">WHITESPACE</span><span class="p">:</span><span class="n">example</span><span class="o">/</span><span class="n">file</span><span class="o">.</span><span class="n">html</span><span class="p">:</span><span class="mi">128</span>
</pre></div>
</div>
</section>
</section>


          </div>
          
        </div>
      </div>
      <div class="sphinxsidebar" role="navigation" aria-label="main navigation">
        <div class="sphinxsidebarwrapper">
<h1 class="logo"><a href="../index.html">web-platform-tests</a></h1>








<h3>Navigation</h3>
<ul class="current">
<li class="toctree-l1"><a class="reference internal" href="../test-suite-design.html">Test Suite Design</a></li>
<li class="toctree-l1"><a class="reference internal" href="../intro-video-transcript.html">“Introduction to WPT” video transcript</a></li>
<li class="toctree-l1"><a class="reference internal" href="../running-tests/index.html">Running Tests</a></li>
<li class="toctree-l1 current"><a class="reference internal" href="index.html">Writing Tests</a><ul class="current">
<li class="toctree-l2"><a class="reference internal" href="index.html#test-types">Test Types</a></li>
<li class="toctree-l2"><a class="reference internal" href="index.html#submitting-tests">Submitting Tests</a></li>
<li class="toctree-l2 current"><a class="reference internal" href="index.html#table-of-contents">Table of Contents</a></li>
</ul>
</li>
<li class="toctree-l1"><a class="reference internal" href="../reviewing-tests/index.html">Reviewing Tests</a></li>
<li class="toctree-l1"><a class="reference internal" href="../admin/index.html">Project Administration</a></li>
</ul>

<div class="relations">
<h3>Related Topics</h3>
<ul>
  <li><a href="../index.html">Documentation overview</a><ul>
  <li><a href="index.html">Writing Tests</a><ul>
      <li>Previous: <a href="submission-process.html" title="previous chapter">Submitting Tests</a></li>
      <li>Next: <a href="ahem.html" title="next chapter">The Ahem Font</a></li>
  </ul></li>
  </ul></li>
</ul>
</div>
<div id="searchbox" style="display: none" role="search">
  <h3 id="searchlabel">Quick search</h3>
    <div class="searchformwrapper">
    <form class="search" action="../search.html" method="get">
      <input type="text" name="q" aria-labelledby="searchlabel" autocomplete="off" autocorrect="off" autocapitalize="off" spellcheck="false"/>
      <input type="submit" value="Go" />
    </form>
    </div>
</div>
<script>document.getElementById('searchbox').style.display = "block"</script>








        </div>
      </div>
      <div class="clearer"></div>
    </div>
    <div class="footer">
      &#169;2019, wpt contributors.
      
      |
      Powered by <a href="https://www.sphinx-doc.org/">Sphinx 7.2.6</a>
      &amp; <a href="https://alabaster.readthedocs.io">Alabaster 0.7.16</a>
      
      |
      <a href="../_sources/writing-tests/lint-tool.md.txt"
          rel="nofollow">Page source</a>
    </div>

    

    
  </body>
</html>
